# perf(lockfiles): Eagerly parse berry resolutions