--- a/crates/turborepo-lockfiles/src/berry/resolution.rs
+++ b/crates/turborepo-lockfiles/src/berry/resolution.rs
@@ -22,21 +22,24 @@ pub enum Error {
     UnexpectedEOI,
     #[error("unexpected token")]
     UnexpectedToken(Rule),
+    #[error("invalid identifier used as specifier: {0}")]
+    InvalidSpecifier(#[from] super::identifiers::Error),
 }
 
 /// A resolution that can appear in the resolutions field of the top level
 /// package.json
-#[derive(Debug, PartialEq, Clone, Copy, Eq, Default, PartialOrd, Ord, Hash)]
+#[derive(Debug, PartialEq, Clone, Eq, PartialOrd, Ord, Hash)]
 pub struct Resolution<'a> {
     from: Option<Specifier<'a>>,
     descriptor: Specifier<'a>,
 }
 
 // This is essentially an Ident with an optional semver range
-#[derive(Debug, PartialEq, Clone, Copy, Eq, Default, PartialOrd, Ord, Hash)]
+#[derive(Debug, PartialEq, Clone, Eq, PartialOrd, Ord, Hash)]
 struct Specifier<'a> {
     full_name: &'a str,
     description: Option<&'a str>,
+    ident: Ident<'a>,
 }
 
 #[derive(Parser)]
@@ -82,10 +85,8 @@ fn parse_specifier(specifier: Pair<'_, Rule>) -> Result<Option<Specifier>, Error
             let mut parts = specifier.into_inner();
             let full_name = parts.next().ok_or(Error::UnexpectedEOI)?.as_str();
             let description = parts.next().map(|p| p.as_str());
-            Ok(Some(Specifier {
-                full_name,
-                description,
-            }))
+            let spec = Specifier::new(full_name, description)?;
+            Ok(Some(spec))
         }
         Rule::EOI => Ok(None),
         _ => Err(Error::UnexpectedToken(specifier.as_rule())),
@@ -108,50 +109,40 @@ impl<'a> Resolution<'a> {
         if let Some(from) = &self.from {
             let from_ident = from.ident();
             // If the from doesn't match the locator we skip
-            if from_ident != locator.ident {
+            if from_ident != &locator.ident {
                 return None;
             }
 
-            let mut from_locator = Locator {
-                ident: from_ident,
-                reference: from
-                    .description
-                    .map_or_else(|| locator.reference.to_string(), |desc| desc.to_string())
-                    .into(),
-            };
-
-            // we now insert the default protocol if one isn't present
-            if Version::parse(&from_locator.reference).is_ok()
-                || tag_regex().is_match(&from_locator.reference)
-            {
-                let reference = from_locator.reference.to_mut();
-                reference.insert_str(0, "npm:");
-            }
-
-            // If the normalized from locator doesn't match the package we're currently
-            // processing, we skip
-            if &from_locator != locator {
-                return None;
+            // TODO flatten this out
+            if let Some(desc) = from.description {
+                match Version::parse(&desc).is_ok() || tag_regex().is_match(&desc) {
+                    true => {
+                        // we check if npm:{desc} is equiv, possible to avoid alloc
+                        if let Some(r) = locator.reference.strip_prefix("npm:") {
+                            if desc != r {
+                                return None;
+                            }
+                        }
+                    }
+                    false => {
+                        if desc != locator.reference {
+                            return None;
+                        }
+                    }
+                }
             }
         }
 
         // Note: berry parses this as a locator even though it's an ident
         let resolution_ident = self.descriptor.ident();
-        if resolution_ident != dependency.ident {
+        if resolution_ident != &dependency.ident {
             return None;
         }
 
-        let resolution_descriptor = Descriptor {
-            ident: resolution_ident,
-            range: self
-                .descriptor
-                .description
-                .map_or_else(|| dependency.range.to_string(), |range| range.to_string())
-                .into(),
-        };
-
-        if &resolution_descriptor != dependency {
-            return None;
+        if let Some(resolution_range) = self.descriptor.description {
+            if resolution_range != dependency.range {
+                return None;
+            }
         }
 
         // We have a match an we now override the dependency
@@ -180,8 +171,18 @@ impl<'a> Resolution<'a> {
 }
 
 impl<'a> Specifier<'a> {
-    fn ident(&self) -> Ident<'a> {
-        Ident::try_from(self.full_name).expect("Invalid identifier in resolution")
+    pub fn new(full_name: &'a str, description: Option<&'a str>) -> Result<Specifier<'a>, Error> {
+        let ident = Ident::try_from(full_name)?;
+
+        Ok(Specifier {
+            full_name,
+            description,
+            ident,
+        })
+    }
+
+    pub fn ident(&self) -> &Ident<'a> {
+        &self.ident
     }
 }
 
@@ -247,10 +248,7 @@ mod test {
             parse_resolution("relay-compiler").unwrap(),
             Resolution {
                 from: None,
-                descriptor: Specifier {
-                    full_name: "relay-compiler",
-                    description: None
-                }
+                descriptor: Specifier::new("relay-compiler", None).unwrap()
             }
         )
     }
@@ -261,10 +259,7 @@ mod test {
             parse_resolution("@babel/core").unwrap(),
             Resolution {
                 from: None,
-                descriptor: Specifier {
-                    full_name: "@babel/core",
-                    description: None
-                }
+                descriptor: Specifier::new("@babel/core", None,).unwrap()
             }
         )
     }
@@ -274,28 +269,16 @@ mod test {
         assert_eq!(
             parse_resolution("webpack/memory-fs").unwrap(),
             Resolution {
-                from: Some(Specifier {
-                    full_name: "webpack",
-                    description: None
-                }),
-                descriptor: Specifier {
-                    full_name: "memory-fs",
-                    description: None
-                }
+                from: Some(Specifier::new("webpack", None).unwrap()),
+                descriptor: Specifier::new("memory-fs", None).unwrap()
             }
         );
 
         assert_eq!(
             parse_resolution("is-even/is-odd").unwrap(),
             Resolution {
-                from: Some(Specifier {
-                    full_name: "is-even",
-                    description: None
-                }),
-                descriptor: Specifier {
-                    full_name: "is-odd",
-                    description: None
-                }
+                from: Some(Specifier::new("is-even", None).unwrap()),
+                descriptor: Specifier::new("is-odd", None).unwrap()
             }
         );
     }
@@ -305,14 +288,8 @@ mod test {
         assert_eq!(
             parse_resolution("@babel/core@npm:7.0.0/@babel/generator").unwrap(),
             Resolution {
-                from: Some(Specifier {
-                    full_name: "@babel/core",
-                    description: Some("npm:7.0.0"),
-                }),
-                descriptor: Specifier {
-                    full_name: "@babel/generator",
-                    description: None
-                }
+                from: Some(Specifier::new("@babel/core", Some("npm:7.0.0"),).unwrap()),
+                descriptor: Specifier::new("@babel/generator", None).unwrap()
             }
         )
     }