… crates/turborepo-lockfiles/examples/yarn.lock